[GSoC18] Refactored Inflation for FormulaBricks
 - added a map that contains mappings between text views
   and BrickFields.
 - these are now automatically set up and updated.
 - refactored Brick constructors.
 - mitigated inheritance hell for if-else logic.
 - refactored FormulaEditorFragment: simplified and removed
   unused/unnecessary code.
 - BrickField highlighting no longer breaks layout
   (is a little more subtle though).

+++ catroid/src/main/java/org/catrobat/catroid/content/Sprite.java
@@ -604,7 +604,7 @@ public class Sprite implements Serializable, Cloneable {
 
 	public class PenConfiguration {
 		public boolean penDown = false;
-		public float penSize = BrickValues.PEN_SIZE;
+		public double penSize = BrickValues.PEN_SIZE;
 		public Color penColor = BrickValues.PEN_COLOR;
 		public PointF previousPoint = null;
 		public boolean stamp = false;
